index.html , css folder, images folder
added text changes to headers and removed a hr tag.   Added a css file for portal assignment.  Resided images pictures of Martin. Resized headers and colored the main heading.  Resized text and changed the percentage of Media all for desktop to 98% size.   Resized the margins and padding for Nav Responsive container to fit across entire page better.  Tested the html validation and css.  The css has a font awesome a auto error.  
    The Responsive @media queries don't work with the changing of devices in the inspection view port.  I need help here!