Dreamhost did not work good with this code!
I have a reason why Dream host sucked file in the Web120 folder section.  Its shows an image of proof dreamhost uploaded the wrong file in my Web120/images/ folder. 


I Feel the Ten times I contacted Dreamhost for help to start my website up on the filezilla as gone too far with the trouble for this assignment as well.