This png is for the portal assignment,.
That is the index.html at this time.  The images folder does use the desktop.jpg image.  The file that is rendering is not that image its phone.jpg.  So the Proof is in that the Dreamhost glitchs with ROM Shadowing are at a point of trouble for Dreamhost.  My assignments can not render in time.  The next day they just render to a different web page.  The wrong idea for a 100% on any assignment.